requirements.txt
kafka-python
pandas
pyarrow
delta-spark
psycopg2-binary
boto3 # For AWS S3 interactions from Python
geopy